Jax version of MGE Sersic profile
This adds the MGE code needed for the Sersic profile.
